Feature loc7 (#59)
* Update

* Add default altitude & indicated airspeed import options

* Add default country combobox to CSV import options

* Default country for Little Navmap CSV import

* Use the CSVParser for the location migration

* Use CSVParser, reserve memory

* Reserve memory, use move assignment

* Add missing include

* Add engine event to location (WIP)

- TODO: Provide event (FK violation)

* Fix compilation (Linux)

* Use Export::formatNumber, Unit::formatCoordinate

* Always use decimal point for coordinates

* Add default "KEEP" engine event

* Add engine event to location table

* Send engine start/stop events when teleporting

* Location table layout changed

* Migrate locations

* Sky Dolly CSV import refactoring (WIP)

* Import/exort engine event (locations)

* Sky Dolly CSV import refactoring (WIP 2)

* Apply common time offset to all imported CSV timestamps

- Based on first position sample timestamp

* Make export methods const

* Make export locations interface const

* Direct member access

* Pre-allocate memory

* Use enum struct for indices

* Rename variables

* Replace backup period combobox with enumeration combobox

- WIP
- Need to remove "Next time" when showing actual backup dialog
- Properly persist settings

* General performance improvements

- Based on code clang-tidy performance analysis
- Use of std::move(), "const Tpye &", noexcept and others

* Use firstLine instead of row counter

* Use the CSVParser for FlightRecorder import

* Properly read backup period from metadata

* Fix compilation (Linux, gcc)

* Fix compilation (on case-sensitive file systems)

* Fix compilation (Linux, gcc)

* Shorten symbolicId to symId

* Shorten toUnderlyingType to underly

* Fix a memory leak when creating a Version with given major, minor and
patch numbers

* Ignore backup periods depending on context

- Context: settings or backup dialog

* Update

* Ignore the "Now" backup period in the backup dialog

- TODO: What is the "Now" backup period good for?

* Fix compilation (Windows)

* Renumerate backup periods

* Add Como Hidroport Airport (LILY)

+++ src/Kernel/include/Kernel/SampleRate.h
@@ -161,7 +161,7 @@ namespace SampleRate
         if (sampleRateValue != AutoValue) {
             return static_cast<int>(1000.0 / sampleRateValue);
         } else {
-            return 1000.0 / DefaultAutoSampleRate;
+            return static_cast<int>(1000.0 / DefaultAutoSampleRate);
         }
     }
